feat: Move scripts to folders
