+++ src/menus/scrambler_menu.py
